backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyYAML>=6.0.1
python-dotenv>=1.0.1

openai>=1.40.0
numpy>=2.0.0
faiss-cpu>=1.13.0

